Add files via upload
Keeping the Description Short:

The dataset contains the conversations of the people. To understand the conversation and finding the summary is a time taking. So, this Model can reflect the summary and thus it saves a lot of time to analyze what they are talking about. This Model can give the summarization very quickly and very efficiently.